first

+++ rolling/aggregation/count_test.go
@@ -14,7 +14,7 @@ func TestCount(t *testing.T) {
 			ExpectedBow: func() bow.Bow {
 				b, err := bow.NewBow(
 					bow.NewSeries("time", bow.Int64, []int64{}, nil),
-					bow.NewSeries("value", bow.Int64, []float64{}, nil),
+					bow.NewSeries("value", bow.Int64, []int64{}, nil),
 				)
 				assert.NoError(t, err)
 				return b
